reorg datamodels
